+++ src/index.ts
@@ -2,3 +2,4 @@ export * as base64 from './base64';
 export * as base64Url from './base64Url';
 export * as pkce from './pkce';
 export { default as sha256 } from './sha256';
+export { default as sha256Async } from './sha256Async';