+++ src/isBrowser/index.spec.ts
@@ -1,0 +1,8 @@
+import isBrowser from './index';
+
+describe('sha256Async', () => {
+  it('should check if is browser', async () => {
+    const actual = isBrowser();
+    expect(actual).toBeFalsy();
+  });
+});