+++ src/isBrowser/index.ts
@@ -1,0 +1,5 @@
+const isBrowser = (): boolean => {
+  return process.release.name !== 'node';
+};
+
+export default isBrowser;